# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the signed-in user profile",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the signed-in user profile",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedAccountDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Permanently delete my account and all of its data",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/onboarding/info": {
      "post": {
        "operationId": "UsersController_onboardingInfo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingInfoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Account setup step 1 — name, gender, age (country is server-derived)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/onboarding/permissions": {
      "post": {
        "operationId": "UsersController_onboardingPermissions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingPermissionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Account setup step 2 — notification permission",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/onboarding/interests": {
      "post": {
        "operationId": "UsersController_onboardingInterests",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingInterestsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Account setup step 3 — interests",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/onboarding/goal": {
      "post": {
        "operationId": "UsersController_onboardingGoal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingGoalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Account setup step 4 — goal, finishes onboarding",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/notification-preferences": {
      "patch": {
        "operationId": "UsersController_updateNotificationPreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Turn push categories on or off, and set quiet hours opt-in",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/device": {
      "post": {
        "operationId": "UsersController_registerDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register or refresh this device — fraud signals and push notification targeting",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coinzu API",
    "description": "Coinzu backend API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "device_id": {
            "type": "string",
            "example": "b7e2b6b0-1a2b-4c3d-9e4f-5a6b7c8d9e0f",
            "description": "The app's persistent per-install device id."
          },
          "platform_type": {
            "type": "string",
            "example": "android",
            "enum": [
              "ios",
              "android",
              "web"
            ],
            "description": "Falls back to User-Agent sniffing when omitted."
          },
          "push_token": {
            "type": "string",
            "example": "fcm-or-apns-token",
            "description": "Push notification token. Send again whenever it rotates."
          },
          "hardware_id": {
            "type": "string",
            "example": "ff8a1c2d3e4f5a6b",
            "description": "Reinstall-surviving device id (SSAID on Android, IDFV on iOS). Server links accounts sharing it. Fraud signal — server uses it, never returns it."
          },
          "sim_country_code": {
            "type": "string",
            "example": "IN",
            "description": "SIM ISO 3166-1 alpha-2 country. Null on wifi-only / eSIM / web."
          },
          "carrier": {
            "type": "string",
            "example": "Airtel",
            "description": "SIM carrier name."
          },
          "mcc_mnc": {
            "type": "string",
            "example": "404-45",
            "description": "Mobile country code + network code, \"MCC-MNC\"."
          },
          "is_emulator": {
            "type": "boolean",
            "example": false,
            "description": "True when the app is running on an emulator, not real hardware."
          },
          "is_rooted": {
            "type": "boolean",
            "example": false,
            "description": "True when the device appears rooted / jailbroken."
          },
          "device_info": {
            "type": "object",
            "example": {
              "app_version": "1.4.2",
              "os_version": "17.4",
              "model": "Pixel 8",
              "brand": "Google",
              "locale": "en-US",
              "timezone": "Asia/Kolkata"
            },
            "description": "Free-form device metadata. Any JSON object — add fields as the app needs them, no migration required."
          }
        },
        "required": [
          "device_id"
        ]
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "cz_user_id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+919875643266"
          },
          "google_id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true,
            "example": "Ada Lovelace"
          },
          "gender": {
            "type": "string",
            "nullable": true,
            "example": "female"
          },
          "age_range": {
            "type": "string",
            "nullable": true,
            "enum": [
              "18-24",
              "25-34",
              "35-44",
              "45-54+"
            ]
          },
          "country": {
            "type": "string",
            "nullable": true,
            "example": "GB"
          },
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "interests": {
            "example": [
              "action",
              "puzzle"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "primary_goal": {
            "type": "string",
            "nullable": true
          },
          "referral_code": {
            "type": "string",
            "example": "ADA2026"
          },
          "referred_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "tier": {
            "type": "string",
            "enum": [
              "silver",
              "gold",
              "platinum",
              "diamond"
            ]
          },
          "kyc_status": {
            "type": "string",
            "enum": [
              "none",
              "pending",
              "verified",
              "rejected",
              "manual_review"
            ]
          },
          "profile_completion_pct": {
            "type": "number",
            "example": 71
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "banned",
              "deleted"
            ]
          },
          "onboarding_completed": {
            "type": "boolean"
          },
          "notifications_enabled": {
            "type": "boolean"
          },
          "email_verified_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "phone_verified_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "cz_user_id",
          "email",
          "phone",
          "google_id",
          "name",
          "gender",
          "age_range",
          "country",
          "avatar_url",
          "interests",
          "primary_goal",
          "referral_code",
          "referred_by",
          "tier",
          "kyc_status",
          "profile_completion_pct",
          "role",
          "status",
          "onboarding_completed",
          "notifications_enabled",
          "email_verified_at",
          "phone_verified_at",
          "last_login_at",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "gender": {
            "type": "string",
            "example": "female"
          },
          "age_range": {
            "type": "string",
            "example": "25-34",
            "enum": [
              "18-24",
              "25-34",
              "35-44",
              "45-54+"
            ]
          },
          "country": {
            "type": "string",
            "example": "GB",
            "description": "ISO 3166-1 alpha-2."
          },
          "avatar_url": {
            "type": "string",
            "example": "https://cdn.coinzu.app/avatars/a.png"
          },
          "interests": {
            "example": [
              "gaming",
              "shopping"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phone": {
            "type": "string",
            "example": "+919875643266",
            "description": "E.164 format, digits with an optional leading +."
          }
        }
      },
      "DeletedAccountDto": {
        "type": "object",
        "properties": {
          "cz_user_id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          }
        },
        "required": [
          "cz_user_id",
          "email"
        ]
      },
      "OnboardingInfoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "gender": {
            "type": "string",
            "example": "female"
          },
          "age_range": {
            "type": "string",
            "example": "25-34",
            "enum": [
              "18-24",
              "25-34",
              "35-44",
              "45-54+"
            ]
          }
        },
        "required": [
          "name",
          "gender",
          "age_range"
        ]
      },
      "OnboardingPermissionsDto": {
        "type": "object",
        "properties": {
          "notifications_enabled": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "notifications_enabled"
        ]
      },
      "OnboardingInterestsDto": {
        "type": "object",
        "properties": {
          "interests": {
            "example": [
              "gaming",
              "shopping",
              "travel"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "interests"
        ]
      },
      "OnboardingGoalDto": {
        "type": "object",
        "properties": {
          "primary_goal": {
            "type": "string",
            "example": "earn_side_income"
          }
        },
        "required": [
          "primary_goal"
        ]
      },
      "UpdateNotificationPreferencesDto": {
        "type": "object",
        "properties": {
          "notifications_enabled": {
            "type": "boolean",
            "description": "Master switch for push on this account."
          },
          "announcement": {
            "type": "boolean",
            "description": "Product news and general announcements."
          },
          "promotion": {
            "type": "boolean",
            "description": "Offers, bonuses and campaigns."
          },
          "reward": {
            "type": "boolean",
            "description": "Streaks, achievements and reward nudges."
          },
          "quiet_hours": {
            "type": "boolean",
            "description": "Hold marketing back overnight. On by default."
          }
        }
      },
...
    }
  }
}